requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pymupdf>=1.23.5
beautifulsoup4>=4.12.2
requests>=2.31.0
gradio>=3.41.0,<4.0.0
fastapi>=0.94.0,<0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic==1.10.8
lxml>=4.9.0
certifi>=2023.7.22
urllib3==1.26.15
xmltodict>=0.13.0
